requirements.txt
Django==5.0
djangorestframework
firebase-admin
python-dotenv
redis
